Struct sim2d::application::GlfwWindow

pub struct GlfwWindow { /* private fields */ }

All resources required for running a single-windowed GLFW application which renders graphics using Vulkan.

GlfwWindow derefs as a raw GLFW window handle so application state can configure the window however is convenient.

Implementations

impl GlfwWindow

pub fn new( window_title: impl AsRef<str> ) -> Result<(Self, Receiver<(f64, WindowEvent)>)>

Create a new GLFW window.

The window starts in “windowed” mode and can be toggled into fullscreen or resized by the application.

Params

window_title - The title shown on the window’s top bar.

pub unsafe fn create_render_device(&self) -> Result<Arc<RenderDevice>>

Create a render device for the application.

Safety

The application is responsible for synchronizing access to all Vulkan resources and destroying the render device at exit.