sim2d

Module graphics_2d

G2 is the 2D graphics API. It support drawing arbitrary shapes which typically change every frame.

Re-exports

pub use self::mesh::GeometryMesh;

pub use self::mesh::Vertex;

pub use self::texture::Texture;

pub use self::texture::TextureAtlas;

pub use self::texture::TextureLoader;

Modules

dynamic_buffer 🔒

frame_constants 🔒

material 🔒

mesh 🔒

texture 🔒

utility 🔒

Structs

DrawParams 🔒

FrameDraw 🔒

All of the resources required to assemble draw commands for a frame.

Graphics2D

The 2D Graphics entrypoint.

MeshTransform 🔒

Constants

INITIAL_CAPACITY 🔒